requirements.txt
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
boto3==1.12.12
botocore==1.15.12
cachetools==4.0.0
cchardet==2.1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
datapackage==1.11.1
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
docutils==0.15.2
et-xmlfile==1.0.1
Flask==1.1.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gspread==3.2.0
httplib2==0.17.0
idna==2.9
ijson==2.6.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.5
jsonlines==1.2.0
jsonpointer==2.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
linear-tsv==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.3
packaging==20.1
Pillow==7.1.1
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
pyxero==0.9.2
requests==2.23.0
requests-oauthlib==1.3.0
rfc3986==1.3.2
rsa==4.0
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.1
tableschema==1.14.0
tabulator==1.35.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.11.2
xlrd==1.2.0
